Improvements to PILImage based on improvements in underlying image code.
